+++ app/src/main/java/vn/asiantech/internship/ultil/DepthPageTransformer.java
@@ -1,4 +1,4 @@
-package vn.asiantech.internship.ultils;
+package vn.asiantech.internship.ultil;
 
 import android.support.v4.view.ViewPager;
 import android.view.View;
