+++ app/src/main/java/vn/asiantech/internship/ultil/MyCustomTab.java
@@ -1,4 +1,4 @@
-package vn.asiantech.internship.ultils;
+package vn.asiantech.internship.ultil;
 
 import android.content.Context;
 import android.graphics.Canvas;
